+++ src/ts/types/index.d.ts
@@ -176,7 +176,7 @@ interface IPreview {
 }
 
 interface IPreviewOptions {
-    className?: string;
+    theme?: "classic" | "dark";
     customEmoji?: { [key: string]: string };
     lang?: (keyof II18nLang);
     emojiPath?: string;